requirements.txt
Flask
Flask-SocketIO
eventlet
requests
PyJWT
flask-sock
